+++ docs/CHANGELOG.md
@@ -1,3 +1,11 @@
+
+## Fixes
+
+- Fixed an issue where if the user accidentally hit 'r' the document would rotate and people
+  couldn't figure out how to revert.
+
+# 1.60.11
+
 ## Features
 
 - The sidebar now doesn't show the color for the buttons unless you hover over them.  The colors